webhook-middleware/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
pillow==8.4.0
Flask==2.0.2
waitress==2.1.2
zeep==4.1.0
